Fix IndexError when collating chat history (#195)
* Fix IndexError when collating chat history

This fixes a bug which causes the exception:
```
ERROR:root:Exception in /api/conversation/custom | list index out of range
Traceback (most recent call last):
  File "/workspaces/chat-with-your-data-solution-accelerator/code/app/app.py", line 283, in conversation_custom
    chat_history.append((user_assistant_messages[i]['content'],user_assistant_messages[i+1]['content']))
```

This is caused when there has been an error providing a response, and
the latest message in the history is from a user, rather than the
assitant. Our code assumes a user message is always followed by an
assistant message.

This change removes that assumption and explitely retreives the role for
each message when collating the chat history.

Required by https://github.com/Azure-Samples/chat-with-your-data-solution-accelerator/issues/114

* Add python formatter to dev container

* Add tests for conversation custom

- Extract some elements to dedicated function to allow mocking

* Switch to black formatter to align with precommit hook

* Add test to cover error scenario when message index is out of range

* Add dependencies required for running app tests

---------

Co-authored-by: Ross Smith <ross-p-smith@users.noreply.github.com>